+++ lib/notion/blocks/block.dart
@@ -5,7 +5,8 @@ import 'package:notion_api/utils/utils.dart';
 /// A base representation of any Notion block object.
 class Block extends BaseFields {
   /// The type of object. Always Block for this.
-  ObjectTypes object = ObjectTypes.Block;
+  @override
+  final ObjectTypes object = ObjectTypes.Block;
 
   /// The block id.
   String id = '';
@@ -38,10 +39,10 @@ class Block extends BaseFields {
   bool get isToogle => this.type == BlockTypes.Toggle;
 
   /// Returns true if is a Bulleted block.
-  bool get isBulleted => this.type == BlockTypes.BulletedList;
+  bool get isBulletedItem => this.type == BlockTypes.BulletedListItem;
 
   /// Returns true if is a Numbered block.
-  bool get isNumbered => this.type == BlockTypes.NumberedList;
+  bool get isNumberedItem => this.type == BlockTypes.NumberedListItem;
 
   /// Returns true if is a Child block.
   bool get isChild => this.type == BlockTypes.Child;