+++ test/blocks/todo_test.dart
@@ -1,3 +1,4 @@
+import 'package:notion_api/notion/blocks/paragraph.dart';
 import 'package:notion_api/notion/blocks/todo.dart';
 import 'package:notion_api/notion/general/rich_text.dart';
 import 'package:notion_api/notion/general/types/notion_types.dart';
@@ -18,7 +19,7 @@ void main() {
     });
 
     test('Create an instance with information', () {
-      ToDo todo = ToDo(text: Text('A'), checked: true).add(Text('B'));
+      ToDo todo = ToDo(text: Text('A'), checked: true).addText('B');
 
       expect(todo.checked, true);
       expect(todo.content.length, 2);
@@ -34,21 +35,32 @@ void main() {
           Text('bar'),
         ],
         checked: true,
-      ).add(Text('last'));
+      ).addText('last').addChild(Paragraph(texts: [
+            Text('A'),
+            Text('B'),
+          ]));
 
       expect(todo.checked, true);
       expect(todo.content.length, 4);
       expect(todo.content.first.text, 'first');
       expect(todo.content.last.text, 'last');
+      expect(todo.children.length, 1);
     });
 
     test('Create json from instance', () {
-      Map<String, dynamic> json = ToDo(text: Text('A')).toJson();
+      Map<String, dynamic> json = ToDo(text: Text('A'))
+          .addChild(Paragraph(texts: [
+            Text('A'),
+            Text('B'),
+          ]))
+          .toJson();
 
       expect(json['type'],
           allOf([isNotNull, isNotEmpty, blockTypeToString(BlockTypes.ToDo)]));
       expect(json, contains(blockTypeToString(BlockTypes.ToDo)));
       expect(json[blockTypeToString(BlockTypes.ToDo)]['text'],
+          allOf([isList, isNotEmpty]));
+      expect(json[blockTypeToString(BlockTypes.ToDo)]['children'],
           allOf([isList, isNotEmpty]));
     });
 
@@ -59,6 +71,8 @@ void main() {
           allOf([isNotNull, isNotEmpty, blockTypeToString(BlockTypes.ToDo)]));
       expect(json, contains(blockTypeToString(BlockTypes.ToDo)));
       expect(json[blockTypeToString(BlockTypes.ToDo)]['text'],
+          allOf([isList, isEmpty]));
+      expect(json[blockTypeToString(BlockTypes.ToDo)]['children'],
           allOf([isList, isEmpty]));
     });
   });