+++ CHANGELOG.md
@@ -78,3 +78,31 @@
 * Fix warnings for documentation
 * Improve documentation
 * Add contribution rules
+
+## v1.1.0:
+> Release date: 10/Jul/2021
+* Add more blocks support for `(PATCH): block children` endpoint
+  * `BulletedListItem` block
+  * `NumberedListItem` block
+  * `Toggle` block
+* Add `children` field for blocks:
+  * `BulletedListItem`
+  * `NumberedListItem`
+  * `ToDo`
+  * `Toggle`
+  * `Paragraph`
+* Add methods to manipulate `content` and `children` for blocks:
+  * `addText(String text, {TextAnnotations? annotations})`
+  * `addChild(Block block)`
+  * `addChildren(List<Block> blocks)`
+* Add `Children.withBlocks(List<Block> blocks)` constructor
+* Add `final` for `type` fields to not allow overwrite:
+  * Objects
+  * Blocks
+* Add `BaseClient` class to avoid duplicated code for clients
+* Add `deprecated` annotations for future changes:
+  * Remove `textSeparation` parameter/field
+  * Remove `add(Text text)` function
+  * Remove `texts` getter for `Paragraph`
+  * Remove named parameters for `Children` class
+* Update documentation